fix ExpireTime overflow handling (#41)
*Description of changes:*
Templates on std::chrono::duration are added to make conversions to
ExpireTime simpler:
```cpp
// before
task->waitForCompletion(tasks::ExpireTime::fromNowMillis(200));
// after
using namespace std::chrono_literals;
task->waitForCompletionDelay(200ms);
```
The duration arithmetic also more correctly detects and prevents
overflows by saturating to maximal values.

This branch also provides a temporary hack to the task tests
to allow immediate task polling on the main thread to succeed
by allowing a very small timeout. This does not fix the underlying race.
Also, a convenience method on the StringTable was added to avoid string
allocations when ord is null (could be generalized to if ord is not in
the table, but the null-check avoids taking a lock).


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.