Improve how exceptions are passed. Also reworked symbol table per discussion (#62)
When an exception passes from plugin to Nucleus (or vice-versa), it's
important to capture the "Kind" of exception (this maps to dynamic
exception class in Java), and the message of the exception.

Also reworked symbol table per previous discussion - all symbols are
placed compactly into a data buffer, using obfuscator/deobfuscator
functions to quickly determine if a handle ID is valid. When searching
for a symbol, the same string table is used, with std::map (R-B Tree) to
find an ID from a string.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.