Plugin service name and service config (#67)
This change allows plugins to specify their "service name", and
standardizes how some useful config branches are passed to each plugin.
The service-name is allowed to be changed in the "boostrap" cycle. In
all other cycles, it is passed as one of the parameters.

This change also has some refactoring on how plugin lifecycles are
called in preparation for more complex lifecycle management.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.