Turned on mqtt logging for iot_broker, fixed the data endpoint mixup,… (#75)
Activated logging to STDERR for iot_broker plugin

Fixed the incorrect use of the credEndpoint. It is now connecting to the
dataEndpoint

Removed the use of futures/promises for connecting. Now it retries and
connects correctly.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.