Rewrote callback logic in a way that allows non-static callbacks. (#74)
With this change, the plugin manages callback data that is necessary to
support lambda/functional callbacks. Also opportunity to clean up
callback logic.

A callback is temporarily registered with Nucleus, so that data exists
in both Nucleus and Plugin for each callback. The callback is created in
Plugin, and registered with Nucleus. As much data as is needed is stored
- including any additional parameters. When Nucleus has finished with
callback, it releases the callback so that Plugin can also free
associated data.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.