Add module awareness in local scope - required for logging scope (#88)
With this change, when Nucleus calls into a plugin, it maintains
per-thread information of which module is called. This is used by
logging to track logging configuration, but also used to prevent
callbacks into unloaded plugins.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.