Add logging capability to Nucleus and plugins (#86)
Add Logging feature.
Note that capability also added for deeper tracking of which module is
in context in a call chain, to add to logging capability. That is, when
Nucleus calls into a Plugin, it captures which module is active for the
entire duration of the call chain unless changed.

Logging is configured through Nucleus configuration (per
Greengrass-Java).

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.