Fix running tests out of IDE on Windows (#160)
When running Nucleus unit tests out of IDE on Windows, samples directory
is not found. Added a simple helper to resolve this in a generic way.
The test exposed that exception was thrown before path check, so fixed
code path too.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.