fix(startable): properly get uid/gid from configuration (#183)
*Issue #, if available:*

*Description of changes:*
- properly set user var with returned uid/gid from getUserInfo function.
- bug was that uid/gid was not being updated. UID/GID was always 0/0,
meaning setUserInfo was always being skipped, which forked the child as
same user and group as parent. (not being changing)
- update TODO comment with proper information of why current
implementation is the way it is by skipping as non-root

*Testing* 
- confirmed in test component that effective user and group running
component python script was the ```defaultRunWith posixUser``` in
nucleus configuration
By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.