Rework Serialize to be more generic and not duplicate code (#184)
Moved the "Serialize" model to be more like I had envisioned.
If a class implements Serializable interface, it has a visit(Archive)
method - this may either wrap an archiver or a dearchiver (allowing same
function to be used for all serialization and deserialization without
templates)

Right now only implemented conversion to/from Struct data model - which
is most useful as we can convert to-from JSON and YAML to that data
model. However adapters can be created for the other formats directly -
and will go into a future check-in.

With this change, and ability to convert to/from data model, now able to
clean up the Lifecycle code and fix various bugs along the way.