New LPC pattern (#127)
Changes in Nucleus to implement new LPC/Task/Future pattern:
See
https://quip-amazon.com/gZZoAb8wHyqe/Understanding-LPC-Local-Procedure-Call

Note this is split into multiple commits for purpose of review:
"Remove Examples" - removes a couple of unused examples to reduce code
churn
"Bundled CMake Changes" - all CMake changes together
"Replace scoped handles with Ref-counted handles" - see below
"Changes to introduce Promise and Futures for LPC" - see below
"Apply new LPC to Plugins" - separate out implementation from how it's
used in plugins

Handle Improvements (needed by LPC changes to remove craziness):
1. C++ Plugin abstraction uses std::shared_ptr to manage ref-counting
(to be improved in future). It converts an integer handle to a
ref-counted abstract handle when captured, and calls a release function
when ref-count reaches zero. If a handle needs to be returned by
callback, it's achieved by creating a temporary (stack scope) handle.
1. Module scope is now strictly required else handles breaks. No Stack
scope helpers exist
1. Nucleus handle tables made way more efficient (the inefficiency was
depressing)
1. Removed the "Anchor" concept - no longer needed. Handle logic vastly
simplified, local scope stuff removed

LPC core changes - break down as follows:

Tasks - Tasks are decoupled from everything else, and act as a more
efficient version of std::async - the thread is allocated from a pool
that grows as needed, and shrinks as threads before idle for long enough
time. Emphasis: Tasks are NOT used for LPC calls, they are entirely
decoupled.

LPC - this is essentially a rendezvous between a function provider
(subscription/listener) and a function requester. They are now
synchronous, but optionally return a Future. It's important for the
callbacks to be aware of this and do the right thing. There are three
patterns:
1. Use a function by reference
1. Use a function by name, first handler wins
1. Use a group of functions by name - all handler values returned

Promise/Future - is a means of providing a slot to store a value and
signaling/callbacks when that value has been updated. Either a value or
an error can be set. It also allows a synchronous call to be turned into
an asynchronous call. A few patterns here:
1. Just block and wait for the async future to complete (easiest to
code, blocks current thread, don't do this in Lifecycle or LPC callback
- yes, some code currently blocks in callbacks and needs to be fixed)
1. Create a completion callback. The most interesting is "andThen" which
allows returning of a chained promise that is to be completed when a
dependent future completes.

There is a lot of cleanup to be done, such as renaming, and in
particular, in the plugins, make use of ability to pass exceptions
through the future. The intent of this PR is to get the feature added
and initial port of everything to use new call pattern.

Tested manually with CLI to confirm IPC to MQTT

Tested with the integration tests under `test\ipc`:
```
../../run/tests/test_pubsub.py::test_local_publish[get_data0] Waiting for the response from server...
Connection established with iot core
PASSED
../../run/tests/test_pubsub.py::test_local_publish[get_data1] PASSED
../../run/tests/test_pubsub.py::test_local_publish[get_data2] PASSED
../../run/tests/test_pubsub.py::test_iot_core_publish[get_data0] Waiting for the response from server...
Subscribed with Granted QoS 1
Waiting for the response from server...
PASSED
../../run/tests/test_pubsub.py::test_iot_core_publish[get_data1] Waiting for the response from server...
Subscribed with Granted QoS 1
Waiting for the response from server...
PASSED
../../run/tests/test_pubsub.py::test_iot_core_publish[get_data2] Waiting for the response from server...
Subscribed with Granted QoS 1
Waiting for the response from server...
PASSED
../../run/tests/test_pubsub.py::test_local_subscribe[get_data0] Subscribed to topic my/topic
PASSED
../../run/tests/test_pubsub.py::test_local_subscribe[get_data1] Subscribed to topic my/#
PASSED
../../run/tests/test_pubsub.py::test_iot_core_subscribe[get_data0] Subscribed to topic my/topic with QoS 1
PASSED
../../run/tests/test_pubsub.py::test_iot_core_subscribe[get_data1] Subscribed to topic my/# with QoS 1
PASSEDWaiting for the response from server...
Successfully disconnected mqtt client
```

---------

Co-authored-by: Patrick Cook <cookpate@amazon.com>
Co-authored-by: Anubhav Rawal <rawalexe@amazon.com>