Make sure there really is only one copy of Device SDK and AWS SDK across the project, and fix Windows build (#203)
The existing build was mixing and matching DeviceSDK shared with some
static libraries. In theory, this should be ok, but leads to some
potential edge cases. It kind of built fine on Linux, but blows up in
Windows.

More work is needed as this change pulls more of the SDK in than is
needed, but does move all of the Device SDK into the shared object -
which may be a good thing for size attribution. After this change, the
device SDK library is the largest (~3MB stripped) where all other
plugins are around 200K each.

Tested build/unit tests only - no functional changes in code.