Remove SymbolMapper class
This class is strongly coupled with SymbolValueMap and its functionality
is implemented by public APIs. The effect of the strong coupling can be
seen in !213 where it is impossible to create a SymbolValueMap without
an lvalue reference to a SymbolMapper, and nowhere else is SymbolMapper
directly used by any other interface.