Log manager (#242)
**Issue #, if available:**

**Description of changes:**
Log manager will create a log group and log stream for logs to be
uploaded to, then upload the individual log events from the
greengrass.log file to the Cloudwatch Logs service. It will attempt to
do this every 5 minutes and use credentials vended from TES.

**Why is this change necessary:**
Adds functionality for uploading the greengrass.log file to Cloudwatch.

**How was this change tested:**
Manually tested (with a lower threshold of 30 seconds instead of 5
minutes), by running GG and verifying that the Log Manager plugin is
correctly creating the log group/stream and uploading logs on a regular
cadence. Tested both when the log group doesn't or does exist, and left
running for multiple iterations of log upload periods.

**Any additional information or context required to review the change:**
Does not use the AWS SDK for http calls due to restrictions surrounding
building with it when implementation was started.

**Checklist:**

- [ ] Updated the README if applicable
- [ ] Updated or added new unit tests
- [ ] Updated or added new integration tests
- [ ] Updated or added new end-to-end tests
- [ ] If your code makes a remote network call, it was tested with a
proxy
- [ ] You confirm that the change is backwards compatible

By submitting this pull request, I confirm that you can use, modify,
copy, and
redistribute this contribution, under the terms of your choice.

---------

Co-authored-by: Sameer Zuberi <samzub@amazon.com>
Co-authored-by: Urvashi Jain (urvajain) <urvajain@amazon.com>
Co-authored-by: Patrick Zhang <patrzhan@amazon.com>
Co-authored-by: patrzhan <139926810+patrzhan@users.noreply.github.com>