feat: create lifecycle_ipc plugin and create subscriptions for three … (#276)
…ipc commands

**Issue #, if available:**

**Description of changes:**
Adds the lifecycle_ipc plugin which will implement lifecycle related IPC
commands. The ones that are currently planned to be added are
UpdateState, SubscribeToComponentUpdates, and DeferComponentUpdate. The
subscriptions for these are set up on plugin component start, but
currently do not do anything as they are not implemented. Functions are
set up for future implementation.

**Why is this change necessary:**
Adds the plugin under which certain IPC commands will be implemented in.

**How was this change tested:**
After building/running GG lite, the plugin starts correctly without
error. Added a placeholder testing file for future testing as
implementation is continued.

**Any additional information or context required to review the change:**

**Checklist:**

- [ ] Updated the README if applicable
- [ ] Updated or added new unit tests
- [ ] Updated or added new integration tests
- [ ] Updated or added new end-to-end tests
- [ ] If your code makes a remote network call, it was tested with a
proxy
- [ ] You confirm that the change is backwards compatible

By submitting this pull request, I confirm that you can use, modify,
copy, and
redistribute this contribution, under the terms of your choice.