Reads the deployment doc and populate nucleus config (#280)
**Issue #, if available:**

**Description of changes:**
Will read a deployment doc and from the nucleus's config file and copy
the recipes and artifacts to appropriate folders. Once the coping is
done it will update the nucleus config to have the name and path to
local recipe.

Sample deployment doc  will look like:

```
{
    "serviceModelType": "aws.greengrass#CreateLocalDeploymentRequest",
    "recipeDirectoryPath": "/home/ubuntu/repo/cli/test_deploy/recipes",
    "artifactsDirectoryPath": "/home/ubuntu/repo/cli/test_deploy/artifacts",
    "requestTimestamp": 1714419514479,
    "rootComponentVersionsToAdd": {
        "IoTSmartHomeProvisioner": "1.0.0",
        "sample-componentPython": "1.0.0"
    },
    "requestId": "d2ed2fe3-a418-438f-ffff-2bfff17cc420"
}
```

Fields added to the config.yml:

```
services:
  aws.greengrass.Nucleus-Lite:
    componentType: "NUCLEUS"
    configuration:
.....
      deploymentOnStart:
        path: "/home/ubuntu/repo/cli/depDoc.json"
        lastModifiedTime: 0
.....
```

**Why is this change necessary:**

**How was this change tested:**
After Startup the packages folder was checked to see if it has
components in the correct place and format

**Any additional information or context required to review the change:**

**Checklist:**

- [ ] Updated the README if applicable
- [ ] Updated or added new unit tests
- [ ] Updated or added new integration tests
- [ ] Updated or added new end-to-end tests
- [ ] If your code makes a remote network call, it was tested with a
proxy
- [ ] You confirm that the change is backwards compatible

By submitting this pull request, I confirm that you can use, modify,
copy, and
redistribute this contribution, under the terms of your choice.