Check that deployment is not stale (#296)
**Issue #, if available:**

**Description of changes:**
Check that deployment is not stale when handling a deployment task.

**Why is this change necessary:**
In the future when we convert to a deployment queue, we may end up with
deployments that are stale (timestamp is older than an already-handled
deployment). We want to ignore these deployments so that they do not
override more recent updates from a more recent deployment.

**How was this change tested:**
Added a happy path test for the non-stale case. Include a dummy handler
to pass the task request to that can mark the deployment as successful
for purposes of testing handlers individually (for
ValidateDeploymentHandler and future ones).

**Any additional information or context required to review the change:**
Left a TODO for stale deployment test, which should be revisited when
more of deployments is implemented such that we can add tests with
multiple deployments that save their timestamps to the config on
processing. Current logic to check stale relies on a timestamp that is
currently not set, but will be implemented later during deployment task
completion.

**Checklist:**

- [ ] Updated the README if applicable
- [ ] Updated or added new unit tests
- [ ] Updated or added new integration tests
- [ ] Updated or added new end-to-end tests
- [ ] If your code makes a remote network call, it was tested with a
proxy
- [ ] You confirm that the change is backwards compatible

By submitting this pull request, I confirm that you can use, modify,
copy, and
redistribute this contribution, under the terms of your choice.