Add additional GglByteVec functions
These will be used for common patterns. A lot of `GglByteVec`s are
created from buffers, so a function is useful there. Some places in the
code also need to append a null character or a slash to a string, so
push is useful. Chained methods reduce boilerplate for calling a bunch
of push/appends as they can pass the error forward, and error handling
can be done once at the end.