Prevent subscription callback use after close
Previously the core-bus client subscribe code ran the response callback
without holding the handle's mutex. While this is usually fine as the
handle accesses are protected, this means that the on_response callback
may be running in the subscription thread after another thread has
closed the subscription. This was problematic with the use of the
subscription ctx; since on_response uses the ctx, it must not be running
after closing the subscription, which may release ctx.