Make core bus subscription handling thread-safe
On core bus, subscription handlers must call either ggl_return_err or
ggl_sub_accept. They may only call ggl_respond on a subscription after
calling ggl_sub_accept.

The above requirement was previously not met by subscription handlers,
as they would set up a mechanism for another thread to respond on the
subscription and then accept the subscription, leaving a window in
between those actions that the other thread could respond first. They
did not accept first, as setting up the mechanism could fail and they
would need to respond with an err instead of accepting. While this could
be solved by having splitting the setup mechanism into a fallible part
and a non-fallible part, that would require every subscription handler
to handle this complexity and ensure they ordered the accepting and
setup correctly.

This commit solves the problem generally by having ggl_respond check if
the handler has yet to call ggl_sub_accept, and in that case block until
the handler has done so. With this subscription handlers do not need to
worry about this race condition.

Blocking on the handler in this way requires that subscription handlers
accept or return an error on the subscription before the handler
returns. This precludes deferring the completion of the subscription
request to another thread and continuing with handling further client
requests. This functionality was not being used, and its use would make
the code harder to verify correctness of, so it is fine to lose.