Add static link packages
As the “NVIDIA JETPACK” is based on Ubuntu 22.04. the approach is to use the packages that are available
(without adding some backport package soures list) and just compile the packages from source that are missing statically. 
Use that static libs to link gg-lite static, but also link against shared libs that are available in the correct version.
To do this I did create a container especially for this use case.
Also some CMmakeLists improvements needs to be done to enable that static linking for just those libs that are not available with Ubuntu 22.04.