Use empty handler for SIGPIPE
Setting a signal handler to SIG_IGN is inherited across exec, so child
processes would use it. While we want to ignore it in nucleus processes,
components running under recipe-runner should not get the modified
signal handler. In order to preserve a SIG_IGN set by a parent of a GG
process to its children, we only set the handler if the signal is
SIG_DFN, as setting a handler results in SIG_DFN in children.