Send server error codes over event stream
This commit however is not complete on it's own. Even though the server
is sending the error codes, the client does not yet understand them.
There will be a fast follow up PR for updating the client.

There are a couple of TODOs added to the code which will be removed
once the client starts to process the incoming error codes sent by
the IPC server.